replaced constant to constant.js
