+++ src/components/common/ingredient-details/IngredientDetails.module.css
@@ -1,5 +1,13 @@
+.ingredient_details {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
 .ingredient_headline {
     line-height: 60px;
+    text-align: center;
     width: 100%;
 }
