+++ src/components/constructor-page-components/burger-constructor-bun/constant.js
@@ -1,0 +1,4 @@
+export const positionText = {
+    top: 'верх',
+    bottom: 'низ',
+}
