+++ src/components/constructor-page-components/burger-navigation/BurgerNavigation.jsx
@@ -1,28 +1,14 @@
 import { useSelector, useDispatch } from 'react-redux'
 import { setCurrentTab, setAutoScroll } from '../../../services/store/ingredientTabSlice'
-
 import { Tab } from '@ya.praktikum/react-developer-burger-ui-components'
+
+import { tabList } from './constant'
 
 import styles from './BurgerNavigation.module.css'
 
 export default function BurgerNavigation() {
     const dispatch = useDispatch()
     const currentTab = useSelector(store => store.ingredientTab.currentTab)
-
-    const tabs = [
-        {
-            value: 'bun',
-            title: 'Булки',
-        },
-        {
-            value: 'sauce',
-            title: 'Соусы',
-        },
-        {
-            value: 'main',
-            title: 'Начинки',
-        },
-    ]
 
     function tabHandler(value) {
         dispatch(setCurrentTab(value))
@@ -33,7 +19,7 @@ export default function BurgerNavigation() {
         <div>
             <h1 className='text text_type_main-large'>Соберите бургер</h1>
             <nav className={`${styles.navigation} mt-5`}>
-                {tabs.map(tab => (
+                {tabList.map(tab => (
                     <Tab
                         value={tab.value}
                         active={currentTab === tab.value}
