+++ src/components/constructor-page-components/burger-navigation/constant.js
@@ -1,0 +1,14 @@
+export const tabList = [
+    {
+        value: 'bun',
+        title: 'Булки',
+    },
+    {
+        value: 'sauce',
+        title: 'Соусы',
+    },
+    {
+        value: 'main',
+        title: 'Начинки',
+    },
+]
