+++ src/components/constructor-page-components/ingredient-type-list/constant.js
@@ -1,0 +1,5 @@
+export const headlines = {
+    bun: 'Булки',
+    sauce: 'Соусы',
+    main: 'Начинки',
+}
