+++ src/pages/error/ErrorPage.module.css
@@ -1,0 +1,7 @@
+.error_page  {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 100%;
+}
