+++ src/pages/forgot-password/constant.js
@@ -1,0 +1,19 @@
+const inputs = [
+    {
+        type: 'email',
+        name: 'email',
+        placeholder: 'E-mail',
+    },
+]
+
+const footerData = [
+    {
+        text: 'Вспомнили пароль?',
+        link: {
+            title: 'Войти',
+            href: '/login',
+        },
+    },
+]
+
+export { inputs, footerData }
