+++ src/pages/login/constant.js
@@ -1,0 +1,31 @@
+const inputs = [
+    {
+        type: 'email',
+        name: 'email',
+        placeholder: 'E-mail',
+    },
+    {
+        type: 'password',
+        name: 'password',
+        placeholder: 'Пароль',
+    },
+]
+
+const footerData = [
+    {
+        text: 'Вы - новый пользователь?',
+        link: {
+            title: 'Зарегистрироваться',
+            href: '/register',
+        },
+    },
+    {
+        text: 'Забыли пароль?',
+        link: {
+            title: 'Восстановить пароль',
+            href: '/forgot-password',
+        },
+    },
+]
+
+export { inputs, footerData }
