+++ src/pages/profile/constant.js
@@ -1,0 +1,12 @@
+export const profileNavigationData = [
+    {
+        title: 'Профиль',
+        href: '/profile',
+        message: 'В этом разделе вы можете изменить свои персональные данные',
+    },
+    {
+        title: 'История заказов',
+        href: '/profile/orders',
+        message: 'В этом разделе вы можете посмотреть свою историю заказов',
+    },
+]
